### 2019-09-11 #1723 @Kevin-Mattheus-Moerman
@dphansti thanks for submitting this work.

Before we can consider your work for JOSS you need to address the following points:
- [ ]  JOSS submission require an [OSI approved license](https://opensource.org/licenses). Please choose one and add a LICENSE file to your main folder.
- [ ]  Have a look at our paper [formatting guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Your paper looks like the classic Intro/methods/results/discussion format, at JOSS we have a different format and do not require these sections. Please study the guidelines and reformat your paper accordingly (you can leave the acknowledgements section).
- [ ] Please remove the Source/Information/Contact links from the paper. Such links are added to the paper by default so do not need to appear in the text.
- [ ] Please ensure all references with DOI's have them listed as well

I may have more points once I give your paper a more thorough read, but these are the points for the moment. Let me know if you have questions.
