### 2019-09-17 #1742 @whedon
[REVIEW]: Bedtoolsr: An R package for genomic data analysis and manipulation
**Submitting author:** @dphansti (<a href="http://orcid.org/0000-0003-2123-0051">Douglas Phanstiel</a>)
**Repository:** <a href="https://github.com/PhanstielLab/bedtoolsr.git" target ="_blank">https://github.com/PhanstielLab/bedtoolsr.git</a>
**Version:** v2.29.0-3
**Reviewer:** @nhejazi , @mikmaksi
