### 2019-09-23 #1742 @mikmaksi
Here is my review. Checklist above is also marked. Everything looks good except I could not find a License file. I have one minor suggestion for the installation instructions and to expand on the usage example. Thank you!

### General checks
1. Repository: Is the source code for this software available at the repository url?
   - Yes.
2. License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
   - No. I could not find a license file in the repo.
3. Contribution and authorship: Has the submitting author (@dphansti) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
   - Yes.

### Functionality
1. Installation: Does installation proceed as outlined in the documentation?
   - Yes. I have a minor suggestion about installation instructions. Would the authors consider adding a note about supported operating systems? A user may have R installed on their Windows machine and have bedtools installed on Windows Subsystem for Linux (WSL). bedtoolsr would not work in this case as far as I can tell since R wouldn't be able to call the bedtools binary.
2. Functionality: Have the functional claims of the software been confirmed?
   - I tested bt.intersect, bt.igv, bt.links, bt.closest, bt.bamtobed, bt.coverage and bt.overlap on some sample data and encountered no issues.
3. Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
   - No performance claims to evaluate.

### Documentation
1. A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
   - Yes.
2. Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
   - Yes. See note under the Functionality section.
3. Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
   - The example provided in the README is somewhat basic. An example where multiple bedtools commands are chained together inside a single R script would go a long way to showcasing how bedtoolsr makes it easy to build a whole workflow using genome math in R.
4. Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
   - Yes. Usage instructions are provided for each function through built-in help docs.
5. Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
   - Yes. testthat unit tests are implemented and Travis CI integration is set up.
6. Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
   - Yes.

### Software paper
1. Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
   - Yes.
2. A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
   - Yes. Authors make the case that bedtoolsr allows direct access to all functions in the bedtools package. Users already familiar with bedtools would have no learning curve associated with using bedtoolsr. Other main stregnths of bedtoolsr are ability to pass either file paths or R objects into bedtools functions.
3. State of the field: Do the authors describe how this software compares to other commonly-used packages?
   - Yes. Authors cite Riemondy (IRanges) and Lawrence et al. (valr). bedr package is not mentioned explicitly but is in the Lawrence et al. reference.
4. Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
   - Yes.
5. References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
   - Citations are provided in correct format.
