### 2019-09-29 PhanstielLab/bedtoolsr#2 @nhejazi
Add License
As per the ongoing review of the package for JOSS (https://github.com/openjournals/joss-reviews/issues/1742#event-2641824676), please add a license file. The `DESCRIPTION` file for the package notes the MIT license, so please add both a `LICENSE` file containing the information CRAN will need for the MIT license (including just the year and copyright holder, as described at https://kbroman.org/pkg_primer/pages/licenses.html). Also, please add a file `LICENSE.md` with the full text of the MIT license, including the year and copyright holder information filled in (please make sure to add this second file to the `.Rbuildignore`).
