### 2019-09-29 PhanstielLab/bedtoolsr#3 @nhejazi
Add Contribution Guidelines
As per the ongoing JOSS review, please add contribution guidelines for the package. At minimum, this should include a simple statement about how best to contribute to the package for those interest, placed in the `README`. One example might include adding a "Contributions" section to the `README` (e.g., https://github.com/tlverse/origami#contributions) as well as a `CONTRIBUTING.md` (e.g., https://github.com/tlverse/origami/blob/master/CONTRIBUTING.md) that provides further details.
