### 2019-09-29 #1742 @nhejazi
@dphansti et al., excellent work on this package, looks quite useful! After going through the review checklist, there's just a few items that require your attention before the few outstanding review items can be checked off. Please address the following when convenient:

* The package requires information about an open-source license, see https://github.com/PhanstielLab/bedtoolsr/issues/2
* The package requires contribution guidelines, see https://github.com/PhanstielLab/bedtoolsr/issues/3
* The JOSS paper looks good, but please add DOIs for each item in the bibliography for which this is possible, e.g., for the pybedtools paper in Bioinformatics, the DOI would be 10.1093/bioinformatics/btr539 and the DOI URL https://doi.org/10.1093/bioinformatics/btr539
