### 2019-10-01 0379c0f
add operating system support section
