### 2019-10-01 #1742 @mgymrek
Thanks @nhejazi and @mikmaksi for your helpful reviews.

@dphansti: Please address these comments here. I think they all look minor and addressable.
