### 2019-10-03 f8f144c
Add DOI to applicable references
