+++ paper/paper.bib
@@ -6,7 +6,8 @@
   author  = "Wickham, Hadley",
   journal = "R J.",
   volume  =  3,
-  year    =  2011
+  year    =  2011,
+  doi     = {10.32614/RJ-2011-002}
 }
 
 @ARTICLE{Tippmann2015-yz,
@@ -18,7 +19,8 @@
   pages    = "109--110",
   month    =  jan,
   year     =  2015,
-  language = "en"
+  language = "en",
+  doi      = {10.1038/517109a}
 }
 
 @MISC{noauthor_undated-fe,
@@ -45,7 +47,8 @@
   month    =  jun,
   year     =  2017,
   keywords = "BEDtools; Genomics; Intervals; R; RStudio; reproducibility",
-  language = "en"
+  language = "en",
+  doi      = {10.12688/f1000research.11997.1}
 }
 
 @ARTICLE{Quinlan2010-ak,
@@ -81,7 +84,8 @@
   pages    = "841--842",
   month    =  mar,
   year     =  2010,
-  language = "en"
+  language = "en",
+  doi      = {10.1093/bioinformatics/btq033}
 }
 
 @ARTICLE{Lawrence2013-px,
@@ -108,7 +112,8 @@
   pages    = "e1003118",
   month    =  aug,
   year     =  2013,
-  language = "en"
+  language = "en",
+  doi      = {10.1371/journal.pcbi.1003118}
 }
 
 @ARTICLE{Dale2011-ja,
@@ -134,5 +139,6 @@
   pages    = "3423--3424",
   month    =  dec,
   year     =  2011,
-  language = "en"
+  language = "en",
+  doi      = {10.1093/bioinformatics/btr539}
 }
