### 2019-10-05 90b9261
Added complex example
Code can be found at R/complexExample.R, but also in the README.md
