+++ README.md
@@ -27,7 +27,7 @@ install.packages("/path/to/bedtools_v2.28.0-7", type="source", repos=NULL)
 
 ### Operating System Support
 
-bedtoolsr should work on any system with R and bedtools installed. It has been tested on macOS (version 10.14 "Mojave") and Linux (Ubuntu version 18.04). bedtools is not available for Windows; however, you can either use a virtual machine or [Windows Subsystem for Linux](https://docs.microsoft.com/en-us/windows/wsl/install-win10). In either case, R from the Windows side would not be able to access bedtools running on the Linux side, so R and bedtoolsr would also have to be installed on the Linux side.
+`bedtoolsr` should work on any system with R and bedtools installed. It has been tested on macOS (version 10.14 "Mojave") and Linux (Ubuntu version 18.04). bedtools is not available for Windows; however, you can either use a virtual machine or [Windows Subsystem for Linux](https://docs.microsoft.com/en-us/windows/wsl/install-win10). In either case, R from the Windows side would not be able to access bedtools running on the Linux side, so R and `bedtoolsr` would also have to be installed on the Linux side.
 
 ### Example Usage
 
@@ -40,15 +40,84 @@ B.bed <- data.frame(chrom=c("chr1"), start=15, end=20)
 1 chr1 15 20
 ```
 
+### Complex Example
+
+In this more complex example, loop calls in bedpe format are downloaded from [Phanstiel et. al, 2017](https://www.cell.com/molecular-cell/fulltext/S1097-2765(17)30603-2?_returnURL=https%3A%2F%2Flinkinghub.elsevier.com%2Fretrieve%2Fpii%2FS1097276517306032%3Fshowall%3Dtrue) and CTCF ChIP-seq peak calls are downloaded from [Van Bortle et. al, 2017.](https://genomebiology.biomedcentral.com/articles/10.1186/s13059-017-1310-3) `bedtoolsr` is used to add 5kb on either side of the CTCF peaks with the `bedtoolsr::slop` function before the `bedtoolsr::pairtobed` function computes overlap of CTCF peaks with either both, one, or neither loop anchor from the loop call bedpe file. The total and percentage of loops found in each case is calculated then plotted with `ggplot2`.
+
+```
+## Download and unzip loop calls bedpe file from Phanstiel et. al, 2017
+download.file(url = "https://www.cell.com/cms/10.1016/j.molcel.2017.08.006/attachment/0a5229f1-46bb-4aae-aa42-33651377e633/mmc3.zip",
+              destfile =  "Loops_PMA.txt.zip")
+unzip(zipfile = "Loops_PMA.txt.zip", files = "molcel_6338_Table_S2_Loops_PMA.txt")
+
+
+## Download CTCF Peak bed file from Van Bortle et. al, 2017
+download.file(url = "https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSE96800&format=file&file=GSE96800%5FCTCF%5Fpeak%5FedgeR%5Fraw%2Etxt%2Egz",
+              destfile = "CTCF_peaks.txt.gz")
+
+
+## Read in bed and bedpe files (optional but helfpul for subsequent R operations)
+loopsBedpe  <- read.table(file = "molcel_6338_Table_S2_Loops_PMA.txt", header = T)
+ctcfPeaks   <- read.table(file = gzfile("CTCF_peaks.txt.gz"), header = T)
+
+## Add 5kb up/downstream from each ctcf peak with the bedtoolsr slop function
+ctcfPeaks   <- bedtoolsr::slop(i = ctcfPeaks, g = "data/hg19", b = 5000, header = T)
+
[…]
