### 2019-10-05 46d5018
Update plot with better color order
