### 2019-10-06 5991473
Deleting example R script that breaks build.
