+++ LICENSE.md
@@ -1,4 +1,4 @@
-Copyright (c) 2019, Mayura Patwardhan, Craig Wenger, Doug Phanstiel
+Copyright (c) 2019, Mayura Patwardhan, Craig Wenger, Eric Davis, Doug Phanstiel
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
