### 2019-10-07 #1742 @dphansti
Thank you so much for your quick and helpful reviews.  We believe we have addressed all of the comments and issues.

**Mikmaksi’s comments**

> General Comment 2: I could not find a license file in the repo.

We followed reviewer nhejazi’s instructions below to properly handle licensing for the R package and for the repository.

> Functionality Comment 1: I have a minor suggestion about installation instructions. Would the authors consider adding a note about supported operating systems? A user may have R installed on their Windows machine and have bedtools installed on Windows Subsystem for Linux (WSL). bedtoolsr would not work in this case as far as I can tell since R wouldn't be able to call the bedtools binary.

Great suggestion.  We have added the following text to README.md:
“bedtoolsr should work on any system with R and bedtools installed. It has been tested on macOS (version 10.14 "Mojave") and Linux (Ubuntu version 18.04). bedtools is not available for Windows; however, you can either use a virtual machine or Windows Subsystem for Linux. In either case, R from the Windows side would not be able to access bedtools running on the Linux side, so R and bedtoolsr would also have to be installed on the Linux side.”

> Documentation Comment 3: The example provided in the README is somewhat basic. An example where multiple bedtools commands are chained together inside a single R script would go a long way to showcasing how bedtoolsr makes it easy to build a whole workflow using genome math in R.

We have added a more complex and realistic example to the README.md file.

**nhejazi’s comments**

> The package requires information about an open-source license, see PhanstielLab/bedtoolsr#2

We have added a LICENSE file, referenced by the DESCRIPTION file, with the copyright year and holders, a LICENSE.md with the full license information, and closed the issue. We greatly appreciate the reviewer’s detailed instructions in this regard.

> The package requires contribution guidelines, see PhanstielLab/bedtoolsr#3

We have added this short contribution guidelines section to README.md and closed the issue:
“We welcome user feedback and contributions on this package. If you have a question or a problem, the best approach is to report it is through GitHub's issue tracker. If you want to propose a change to the source code, either to fix a bug or make an improvement, use a pull request.”

> please add DOIs for each item in the bibliography for which this is possible

We have added DOIs for all journal article citations.
