### 2019-10-18 #1742 @mgymrek
@mikmaksi if all looks good can you complete the checklist?

@dphansti this looks great. Can you just address these very minor comments on the proof?
"use to the genomic community" -> "use to the genomics community"
"breakthroughs in genetic research"-> "breakthroughs in genetics research"

After that, can you make a Zenodo archive, ensuring the title and author list match the paper, and report the DOI here? Also please ensure the version listed here (v2.29.0) is correct.
