### 2019-10-18 301dac5
Update paper.md
corrected genomic to genomics and genetic to genetics
