+++ paper/paper.md
@@ -37,7 +37,7 @@ bibliography: paper.bib
 
 # Summary
 
-The sequencing of the human genome and subsequent advances in DNA sequencing technology have created a need for computational tools to analyze and manipulate genomic data sets. The bedtools software suite and the R programming language have emerged as indispensable tools for this purpose but have lacked integration. Here we describe `bedtoolsr`, an R package that provides simple and intuitive access to all bedtools functions from within the R programming environment. We provide several usability enhancements, support compatibility with past and future versions of bedtools, and include unit tests to ensure stability. `bedtoolsr` provides a user-focused, harmonious integration of the bedtools software suite with the R programming language that should be of great use to the genomic community. 
+The sequencing of the human genome and subsequent advances in DNA sequencing technology have created a need for computational tools to analyze and manipulate genomic data sets. The bedtools software suite and the R programming language have emerged as indispensable tools for this purpose but have lacked integration. Here we describe `bedtoolsr`, an R package that provides simple and intuitive access to all bedtools functions from within the R programming environment. We provide several usability enhancements, support compatibility with past and future versions of bedtools, and include unit tests to ensure stability. `bedtoolsr` provides a user-focused, harmonious integration of the bedtools software suite with the R programming language that should be of great use to the genomics community. 
 
 Source Code: https://github.com/PhanstielLab/bedtoolsr.  
 Information: http://phanstiel-lab.med.unc.edu/bedtoolsr.html.  
@@ -46,7 +46,7 @@ Contact: douglas_phanstiel@med.unc.edu
 
 # Introduction
 
-The sequencing of the human genome and subsequent advances in DNA sequencing technology have transformed modern biological research by producing data sets of ever-increasing size and complexity. While these technologies have led to breakthroughs in genetic research, the incredible throughput and breadth of the resulting data have spurred a reliance on computational tools and programming languages to interpret the results. In 2010, Quinlan et al. developed bedtools, a powerful suite of command-line tools for ‘genome arithmetic’ that has become one of the most widely used and indispensable tools for genomic data analysis [@Quinlan2010-ak]. A year later, `pybedtools` extended the features of bedtools to the python programming language [@Dale2011-ja]. During that same time period, the use of the programming language R—with a rich trove of libraries for statistical analysis and data visualization—skyrocketed in the biological sciences [@Tippmann2015-yz]. While several R packages have been developed for bedtools-like genome analysis, their usage and functionality differ significantly from that of bedtools [@Riemondy2017-vz; @Lawrence2013-px]. These differences make them more difficult to use for those who are already familiar with bedtools behavior and lacks some of the capabilities of bedtools.
+The sequencing of the human genome and subsequent advances in DNA sequencing technology have transformed modern biological research by producing data sets of ever-increasing size and complexity. While these technologies have led to breakthroughs in genetics research, the incredible throughput and breadth of the resulting data have spurred a reliance on computational tools and programming languages to interpret the results. In 2010, Quinlan et al. developed bedtools, a powerful suite of command-line tools for ‘genome arithmetic’ that has become one of the most widely used and indispensable tools for genomic data analysis [@Quinlan2010-ak]. A year later, `pybedtools` extended the features of bedtools to the python programming language [@Dale2011-ja]. During that same time period, the use of the programming language R—with a rich trove of libraries for statistical analysis and data visualization—skyrocketed in the biological sciences [@Tippmann2015-yz]. While several R packages have been developed for bedtools-like genome analysis, their usage and functionality differ significantly from that of bedtools [@Riemondy2017-vz; @Lawrence2013-px]. These differences make them more difficult to use for those who are already familiar with bedtools behavior and lacks some of the capabilities of bedtools.
 
[…]
