### 2019-10-18 f784dbf
New release with all changes from review.
