### 2019-10-21 #1742 @dphansti
Thanks @mgymrek. I think we have addressed all of the issues.

1) We made the suggested changes (e.g. genomic -> genomics) in the paper.md file
2) We made a Zenodo archive
3) DOI of the archive is 10.5281/zenodo.3509513
4) The version is v2.29.0-3 is the newest version that has the changes in 1-3 above.

Let me know if we need to do anything else.
