### 2019-10-23 #1742 @mgymrek
Thanks @dphansti. Can you just ensure the Zenodo archive and the paper have the same title? It looks like the Zenodo archive is titled "PhanstielLab/bedtoolsr: bedtoolsr v2.29.0-3".
