### 2019-11-11 #1742 @mgymrek
Thanks for the changes!

@openjournals/joss-eics we're ready to accept this article!
