### 2019-12-05 #1742 @mgymrek
Pinging @openjournals/joss-eics again it looks like this might have been missed.
