### 2019-12-05 c46b8c6
small fix in paper
