### 2019-12-05 #1742 @danielskatz
please merge https://github.com/PhanstielLab/bedtoolsr/pull/4
