### 2019-12-05 6ad0057
fixes to bib
