+++ paper/paper.bib
@@ -4,7 +4,7 @@
 @ARTICLE{Wickham2011-or,
   title   = "testthat: Get Started with Testing",
   author  = "Wickham, Hadley",
-  journal = "R J.",
+  journal = "R Journal",
   volume  =  3,
   year    =  2011,
   doi     = {10.32614/RJ-2011-002}
@@ -117,7 +117,7 @@
 }
 
 @ARTICLE{Dale2011-ja,
-  title    = "Pybedtools: a flexible Python library for manipulating genomic
+  title    = "Pybedtools: a flexible {P}ython library for manipulating genomic
               datasets and annotations",
   author   = "Dale, Ryan K and Pedersen, Brent S and Quinlan, Aaron R",
   abstract = "SUMMARY: pybedtools is a flexible Python software library for
