### 2019-12-05 #1742 @danielskatz
also https://github.com/PhanstielLab/bedtoolsr/pull/5
