### 2019-12-05 #1742 @dphansti
@danielskatz Both have been merged.  Thanks.
